Add --overwrite option
The main motivation behind this change is to prepare for a future
commit. That future commit will make it possible to use xq as a
pre-commit [1] hook. The idea is that the xq pre-commit hook will fail
if any XML files in the repository aren’t formatted the way that xq
would format them.

In order for a pre-commit hook to report a failure, it must do the
following [2]:

> The hook must exit nonzero on failure or modify files.

Before this change, xq did neither of those things. This change makes it
so that xq will modify unformatted files if the --overwrite option is
used. I could have added an option that made xq exit with a nonzero exit
status, but I thought that making xq modify files would be more
convenient and make more sense.

[1]: <https://pre-commit.com>
[2]: <https://pre-commit.com#new-hooks>