Allow multiple files to be passed via the command-line
The main motivation behind this change is prepare for a future commit.
That future commit will make it so that you can use xq as a pre-commit
[1] hook. When pre-commit executes a pre-commit hook, it passes each
file that the hook should be run on as a command-line argument. In order
for xq to work properly as a pre-commit hook, xq needs to support
passing multiple files as command-line arguments.

[1]: <https://pre-commit.com>